+++ CHANGELOG.md
@@ -177,6 +177,7 @@
 - Updated the Pokédex data updated events to always include a Pokemon instance, and optionally a DisguiseData instance.
 - Updated fields in `SpawnNPCPacket` and `SpawnPokemonPacket` to be visible and mutable.
 - Updated `UnvalidatedPlaySoundS2CPacket` to be public instead of internal and made its fields mutable.
+- Added `hideNameTag` field and `HideNPCNameTag` nbt tag to `NPCEntity` to allow hiding the name tag of the NPC.
 
 ### Molang & Datapacks  
 - Added flows for:
@@ -197,6 +198,8 @@
   - For Pokémon:
     - `pokemon.apply(PokemonProperties)`: Applies the given properties to the Pokémon.
     - `pokemon.owner`: Returns the owner of the Pokémon or 0.0 if there is no owner or they are not online.
+- Added NPC field:
+  - `hideNameTag`: Hides the name tag of the NPC.
 
 ## [1.6.0 - The Record Catch Update (December 25th, 2024)](#1-6-0)
 #### "Now that there's a fishing mechanic, the mod is actually good!"